// src/interfaces/loans.ts
export interface ILoan {
  farmer_id: 288;
  loan_product_code: "LP-XWZ48QVG";
  agrodealer_code: null;
  reason: "string";
  created_on: null;
  code: "LON-0QP5R2GPGFJ3";
  requested_amount: 1;
  interest_rate: 0;
  interest: 0;
  fees: 0;
  principal_amount: 1;
  total_amount: 1;
  installments: null;
  installment_amount: null;
  repayment_start_date: null;
  status: string;
  payment_status: "PENDING";
  expected_completion_date: null;
  paid: 0;
  balance: 1;
  product_code: null;
  product_name: "Dairy Loan";
  applicant_name: "Mary Otieno";
  applicant_phone: "0708321651";
  applicant_address: null;
  applicant_email: null;
  applicant_code: "32711464";
  receiver_account_name: null;
  receiver_account_number: null;
  receiver_bank_name: null;
}